do not warn to stderr if one of the logging configs works.
In 0.6.3, if one of the logging configs succeeded, then it was
just used.  If it failed, it failed silently.  This behavior was
expected, and desired.

As the code was here, we tried each log_cfg in the list anyway (possibly
using the last rather than the first) and writing a messgae including
'WARN' to stderr on failure of the log.