restart salt-minion instead of start.
Packages on debian/ubuntu should start on installation.  As a result, if we
want to get config changes we've inserted to be read, we need to restart.

Note one interesting thing here.
upstart considers 'restart' as "restart only if currently running", while
'service' considers restart to be "stop if running, then start".  So the use of
'service' here is important, rather than just 'restart'