check for a marker file by the normal name also
This check is a waste of a stat any time after the migrator module
had run.  As it would take care of moving markers.

However, if the user runs:
 sudo cloud-init modules --mode final

after an upgrade, they'd otherwise run any module that had a '-' in its
name again.  To avoid that, we just return true in that case, and
inform the user how to run the migrator themselves.