change parser.parse 'default_opts' to 'options'
Now, parser.parse specifies options that override any options found,
rather than just being default options.

There could still potentially be a user for default_options, but since we're
not using them anywhere, I've dropped it.  The difference is that in setting up
the root user, we're now insisting that all keys that go in there have the
key_prefix, even if the key content had other options.

I think this is actually the commit that fixes LP: #1136343.