Add the ability to merge with jsonpatch.
Jsonpatch is a new RFC standard for merging
json-like structures which the cloud-init 
cloud-config is one such structure. To use this
in a limited fashion (to start) add the ability
for the cloud-config handler to accept this
content-type and use it as an alternate way to
merge new cloud-config sections into the accumulated
cloud-config.