add util.log_time helper
The reason for this is that more and more things I was wanting to be able to
see how long they took.  This puts that time logic into a single place.  It
also supports (by default) reading from /proc/uptime as the timing mechanism.

While that is almost certainly slower than time.time(), it does give 
millisecond granularity and is not affected by 'ntpdate' having
run in between the two events.