cloudsigma: change default dsmode to 'net'
Previously this had 'local' as the default datasource mode, meaning
that user-data code such as boot hooks and such would not be guaranteed to have
network access.  That would be out of sync with the expectation on other
platforms where the default is 'network up'.

The user can still specify 'dsmode' as local if necessary and the
local datasource will claim itself found.