wait less for the metadata service (by default)
Waiting around for a metadata service in a given datasource means that if its
not there all the subsequent datasources have to wait, and boot is slowed down.

As it is right now, EC2 is the only one that has the right to wait.  In the
past, we had to wait around for the EC2 metadata service.  I really do not want
to extend that courtesy to other cloud platforms.  A network based metadata
service should be up as soon as networking is up.